Move common components
Exporting all the stuff meant there were some conflicts
