+++ packages/notifi-react-card/lib/components/common/ErrorStateCard.tsx
