+++ packages/notifi-react-card/lib/index.ts
@@ -1,5 +1,6 @@
 export * from './components';
 export * from './components/subscription';
+export * from './components/intercom';
 export * from './context';
 export * from './hooks';
 export * from './utils';
